Upgrade shim to 15.7 and add .sbat data section to grub-efi and seloader efi binaries

+++ meta-efi-secure-boot/recipes-bsp/efitools/efitools_1.9.2.bb
@@ -56,14 +56,18 @@ addtask prepare_signing_keys after do_configure before do_compile
 do_prepare_signing_keys[prefuncs] += "check_deploy_keys"
 
 do_install:append() {
-    install -d ${D}${EFI_BOOT_PATH}
-    install -m 0755 ${D}${datadir}/efitools/efi/LockDown.efi ${D}${EFI_BOOT_PATH}
+    if [ x"${UEFI_SB}" = x"1" ]; then
+        install -d ${D}${EFI_BOOT_PATH}
+        install -m 0755 ${D}${datadir}/efitools/efi/LockDown.efi ${D}${EFI_BOOT_PATH}
+    fi
 }
 
 do_deploy() {
     install -d ${DEPLOYDIR}
 
-    install -m 0600 ${D}${EFI_BOOT_PATH}/LockDown.efi "${DEPLOYDIR}"
+    if [ -e ${D}${EFI_BOOT_PATH}/LockDown.efi ] ; then
+        install -m 0600 ${D}${EFI_BOOT_PATH}/LockDown.efi "${DEPLOYDIR}"
+    fi
     if [ -e ${D}${EFI_BOOT_PATH}/LockDown.efi.sig ] ; then
         install -m 0600 ${D}${EFI_BOOT_PATH}/LockDown.efi.sig "${DEPLOYDIR}"
     fi
